% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrmosaic.R
\name{add_clon}
\alias{add_clon}
\title{Returns the putative type of event and clonality for an event}
\usage{
add_clon(lrr = lrr, bdev = bdev)
}
\arguments{
\item{lrr}{logR ratio deviation}

\item{bdev}{bdev deviation}
}
\value{
list with putative clonality and event type ("loh","gain","loss")
}
\description{
Returns the putative type of event and clonality for an event
}

