% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm3.R
\name{adm3_score}
\alias{adm3_score}
\title{Converts read counts and weights to ADM3 scores}
\usage{
adm3_score(d)
}
\arguments{
\item{d}{Vector with log2 ratios}
}
\description{
Converts read counts and weights to ADM3 scores
}

