% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrmosaic.R
\name{annot_ll_and_recalc_sig}
\alias{annot_ll_and_recalc_sig}
\title{Annotate with loglikelihood of the event being a false positive given the number of probes encompassing the event and the amplitude of the GADA signal.}
\usage{
annot_ll_and_recalc_sig(results, resource_dir)
}
\arguments{
\item{results}{output of mr_mosaic_calling_func()}

\item{resource_dir}{path to directory containing the distributions of probecounts and amplitudes of false positive calls.}
}
\value{
input results file where sig.results has the probabilities of an event being a false positive given the number of probes and the amplitude. The joint loglikelihood is given in the "ll" column.
}
\description{
Annotate with loglikelihood of the event being a false positive given the number of probes encompassing the event and the amplitude of the GADA signal.
}

