% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrmosaic.R
\name{annotate_devs}
\alias{annotate_devs}
\title{Computes the rolling median averaged Log2Ratio and B-Allele freq deviation. This method computes the rolling median average of the LogR deviation
in order to use information in homozygous sites to reduce variance.}
\usage{
annotate_devs(mydf = mydf, width = width, by = by)
}
\arguments{
\item{mydf}{Data.frame with Log2ratios and B-allele frequencies}

\item{width}{Width of the window to be used to compute the rolling median average of}

\item{by}{Increment size}
}
\value{
input dataframe with additional logrdev, rollmedianlogr and bdev columns
}
\description{
Computes the rolling median averaged Log2Ratio and B-Allele freq deviation. This method computes the rolling median average of the LogR deviation
in order to use information in homozygous sites to reduce variance.
}

