% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm3.R
\name{breakpoints}
\alias{breakpoints}
\title{Splits the read count distribution into bins}
\usage{
breakpoints(x, max_bin_size = 1000)
}
\arguments{
\item{x}{Vector with log2 ratios}

\item{max_bin_size}{Maximum bin size}
}
\description{
Splits the read count distribution into bins
}

