% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm3.R
\name{calc_adm3_scores}
\alias{calc_adm3_scores}
\title{Generates weights based on mad of log2 ratios within read count range and calculate adm3 scores}
\usage{
calc_adm3_scores(readdata)
}
\arguments{
\item{readdata}{Vector with log2 ratios}
}
\description{
Generates weights based on mad of log2 ratios within read count range and calculate adm3 scores
}

