% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm3.R
\name{calc_log2_ratio}
\alias{calc_log2_ratio}
\title{Generate log2 ratio values for a sample}
\usage{
calc_log2_ratio(data, rdfiles)
}
\arguments{
\item{data}{Output from \code{\link{read_unique_depth_bait_data}}}

\item{rdfiles}{Vector of reference data files (output of cnv_baf_data)}
}
\description{
Generate log2 ratio values for a sample
}

