% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrmosaic.R
\name{compute_stats}
\alias{compute_stats}
\title{Wrapper function to compute probability of significant copy-number and b-allele frequence deviations and combining those using Fisher's method.}
\usage{
compute_stats(mydf = mydf, background = background, window = window,
  by = by, width = width)
}
\arguments{
\item{mydf}{data frame with data points of the sample to be tested}

\item{background}{data frame with data points representing the background distributions (e.g. other chromosome/control sample)}

\item{window}{window size for computing distance-weighted Mann-Whitney U tests for significance of deviations.}

\item{by}{increment size}

\item{width}{window size to compute rolling media average of logR deviations.}
}
\value{
input data.frame with additional "manwh_combined" column containing the test statistic (S) for the Fisher's combined method.
}
\description{
Wrapper function to compute probability of significant copy-number and b-allele frequence deviations and combining those using Fisher's method.
}

