% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.method.R
\name{fisher.method}
\alias{fisher.method}
\title{Fisher's method to combine P-values (adapted from the deprecated MADAM package created by Karl Kugler <karl@eigenlab.net>)}
\usage{
fisher.method(pvals, method = c("fisher"), p.corr = c("bonferroni", "BH",
  "none"), zero.sub = 1e-05, na.rm = FALSE)
}
\arguments{
\item{pvals}{Vector of P-values}

\item{method}{Method used for combining (e.g. "fisher")}

\item{p.corr}{Method for multiple testing correction (e.g. "none","bonferonni", "BH")}

\item{zero.sub}{Substitute zero-valued p-values with this value}

\item{na.rm}{Remove NA values}
}
\description{
Fisher's method to combine P-values (adapted from the deprecated MADAM package created by Karl Kugler <karl@eigenlab.net>)
}

