% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm3.R
\name{generate_correlation_values}
\alias{generate_correlation_values}
\title{Calculates the correlation between samples using read counts at bait regions}
\usage{
generate_correlation_values(data, rdfiles)
}
\arguments{
\item{data}{Output from \code{\link{read_unique_depth_bait_data}}}

\item{rdfiles}{Vector of reference data files (output of cnv_baf_data)}
}
\description{
Calculates the correlation between samples using read counts at bait regions
}

