% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrmosaic.R
\name{get_mdpt_indices}
\alias{get_mdpt_indices}
\title{Get all mid-points of all windows for a vector of positions. Handles windows at the beginning and end of the vector.}
\usage{
get_mdpt_indices(xdf, ws, by)
}
\arguments{
\item{xdf}{vector of positions}

\item{ws}{window size}

\item{by}{increment size}
}
\value{
vector with mid-points of windows
}
\description{
Get all mid-points of all windows for a vector of positions. Handles windows at the beginning and end of the vector.
}

