% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrmosaic.R
\name{merge_nearby_gada_segments}
\alias{merge_nearby_gada_segments}
\title{Merges overlapping segments called by GADA}
\usage{
merge_nearby_gada_segments(mydf)
}
\arguments{
\item{mydf}{data.frame with GADA result segments.}
}
\value{
data.frame with merged segments
}
\description{
Merges overlapping segments called by GADA
}

