% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrmosaic.R
\name{mrmosaic_calling_func}
\alias{mrmosaic_calling_func}
\title{Super function that runs the whole Mr. Mosaic pipeline.}
\usage{
mrmosaic_calling_func(data, t = t, window = 40, width = 5, by = 1)
}
\arguments{
\item{data}{data.frame with the data (i.e. output of reformat_for_mrmosaic())}

\item{t}{T-value for the GADA segmentation algorithm}

\item{window}{window size for computing distance-weighted Mann-Whitney U tests for significance of deviations.}

\item{width}{window size to compute rolling media average of logR deviations.}

\item{by}{increment size}
}
\value{
list containing the used data and called structural variants
}
\description{
Super function that runs the whole Mr. Mosaic pipeline.
}

