% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm3.R
\name{read_unique_depth_bait_data}
\alias{read_unique_depth_bait_data}
\title{Read the read depth information from cnv_baf_data's output}
\usage{
read_unique_depth_bait_data(file, n)
}
\arguments{
\item{file}{Path to file}

\item{n}{Number of rows in file}
}
\value{
Dataframe with read depth information
}
\description{
Read the read depth information from cnv_baf_data's output
}

