% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm3.R
\name{reformat_for_mrmosaic}
\alias{reformat_for_mrmosaic}
\title{Reformats the data.frame with the ADM3 scores to serve as input for the MrMosaic functions}
\usage{
reformat_for_mrmosaic(input_file, adm3_scores)
}
\arguments{
\item{input_file}{Path to *.mrm file generated by the cnv_baf_data program}

\item{adm3_scores}{data.frame with output from calc_adm3_scores()}
}
\value{
data.frame ready to serve as input for the Mr.Mosaic calling function
}
\description{
Reformats the data.frame with the ADM3 scores to serve as input for the MrMosaic functions
}

