% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrmosaic.R
\name{wapplyp}
\alias{wapplyp}
\title{Applies a function using a rolling window approach}
\usage{
wapplyp(x, ws = window, by = NULL, FUN = NULL, ...)
}
\arguments{
\item{x}{vector containing data}

\item{ws}{window size}

\item{by}{increment size (e.g. a value of 1 computes the function on a window around every point)}

\item{FUN}{the function to be applied on every window}

\item{...}{arguments passed to FUN}
}
\value{
output vector
}
\description{
Applies a function using a rolling window approach
}

